requirements.txt

Django==1.8

Pillow==3.2.0
PyYAML==3.11
SQLAlchemy==1.0.12
amqp==1.4.6
anyjson==0.3.3
argparse==1.2.1
bcdoc==0.12.2
billiard==3.3.0.19
boto==2.35.2
botocore==0.84.0
celery==3.1.17
colorama==0.2.5
dj-database-url==0.3.0
dj-static==0.0.6
django-allauth==0.26.1
django-appconf==0.6
django-compress==1.0.1
django-compressor==1.4
django-debug-toolbar==1.2.2

django-pipeline==1.4.5
django-postgrespool==0.3.0
django-storages==1.1.8
docutils==0.12
funcsigs==1.0.2
futures==3.0.5
grunt==0.1.1
gunicorn==19.1.1
jmespath==0.5.0
jsonpickle==0.9.3
kombu==3.0.24
lxml==3.4.0
marrow.interface==1.0.1
marrow.mailer==4.0.0
marrow.util==1.2.3
oauthlib==1.1.2
psycopg2==2.5.4
pyasn1==0.1.7
pyserial==2.7
python-dateutil==2.4.0
python-openid==2.2.5
pytz==2014.10
requests==2.10.0
requests-oauthlib==0.6.2
rsa==3.1.2
simplejson==3.8.2
six==1.9.0
sqlparse==0.1.14
static3==0.6.1
tzlocal==1.2.2
wget==2.2
whitenoise==1.0.6
wsgiref==0.1.2
